change from ASCII to UTF-8 encoding
For the accounts that have no public email address or GPG key (`missed_account`), a list is printed consisting of the Display Name (usually the person's name) and the username (i.e. their Launchpad account). If either includes Unicode, printing fails with `UnicodeEncodeError: 'ascii' codec can't encode character u'\<the character>' in position <location in the string>: ordinal not in range (128)` due to the fact that the default encoding is ASCII rather than UTF-8. This little change will keep that from blowing up and allow for a more complete list of missing accounts.